Split stateless sections into their own files
